continued implementation of the page list component, added more flexibility to the user oft it
